requirements.txt
tornado>=5.0.2
pynetworktables>=2019.0.0
pynetworktables2js>=2019.0.0